requirements.txt
DateTime==4.3
defusedxml==0.7.1
et-xmlfile==1.1.0
Flask==1.0.2
image==1.5.33
lxml==4.5.2
numpy==1.19.2
opencv-python==4.5.2.52
openpyxl==3.0.7
pandas==1.1.3
Pillow==8.2.0
PyPDF2==1.26.0
reportlab==3.5.67
